Concatenation de Main et GUI
